server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.54.0
tavily-python>=0.5.0
pydantic>=2.9.0
python-dotenv>=1.0.1
requests>=2.31.0
sqlmodel>=0.0.14
aiosqlite>=0.19.0
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
pydantic-settings>=2.12.0